@@ -27,15 +27,15 @@ using System;
 using System.Windows.Forms;
 using Weland.Cs.Api.Sample.ImportClient;
 
-namespace OrderImportSample
+namespace Weland.Cs.Api.Sample.ImportClient
 {
-    static class Program
+    internal static class Program
     {
         /// <summary>
         ///  The main entry point for the application.
         /// </summary>
         [STAThread]
-        static void Main()
+        private static void Main()
         {
             Application.SetHighDpiMode(HighDpiMode.SystemAware);
             Application.EnableVisualStyles();
